Login working, changed way collections are loaded from filesystem
The login is now working, hardcoded accounts can be found in
app/scripts/services/auth-service.js

Changed the way collections are retrieved from the filesystem. We might
have to change the way this is done in the future, because loading from
one db into another doesn’t seem like a very efficient way to do this.
If we can get the pagination to work properly in smithers then we don’t
need this intermediate database anymore.